audio_module/pyproject.toml
[project]
name = "audio-module"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.2",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "noisereduce>=3.0.3",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "qwen-tts>=0.1.1",
    "torch>=2.11.0",
    "torchcodec>=0.11.0",
    "torchvision>=0.26.0",
    "uvicorn>=0.42.0",
]

[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
